DiabetesFriend_Server/requirements.txt
fastapi
uvicorn
requests
pymongo
dnspython
pydantic
google-generativeai>=0.8.0
